Further changes on the buildserver

+++ BuildServer/scripts/initialize/default/00-copy.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
-on_error "rm --one-file-system -r '${ROOT%root}'"
+if [[ -v DELETE_ON_FAIL ]]; then on_error "rm --one-file-system -r '${ROOT%root}'"; fi
 
 cp --reflink -r roots/stemgentoo/root/ "${ROOT}"
